Move exercise goals/descriptions to top.
For the most part, the content itself is the same. Also includes
some CSS tweaks I added along the way.

This makes lessons easier to get started with – a student can
get started trying to do something straight away if they're
feeling confident, and it provides some context for the rest of
the content they're reading.