Changed `code` text to more readable color
It might be my lack-luster monitor, but why is there grey on grey text for all the `code` sections of the readme's? I personally changed the color on mine to not have to keep highlighting the text to read it. Also I picked #FFE169 because it is the yellow color used for everything else and matches the color scheme.